Various fixes to hash values representation
Now hash values are 64 bits (which removes the need to re-hash to 32 bits).
Furthermore the code to compute prefixes is now corrected in the case of
odd prefixes.

+++ src/sort.rs
@@ -309,7 +309,7 @@ fn test_radix_sort_hash_pairs() {
     use rand_xoshiro::Xoshiro256PlusPlus;
 
     let mut rng = Xoshiro256PlusPlus::seed_from_u64(12435);
-    let unif = Uniform::new_inclusive(u32::MIN, u32::MAX);
+    let unif = Uniform::new_inclusive(u64::MIN, u64::MAX);
     let v: Vec<(HashValue, u32)> = unif
         .sample_iter(&mut rng)
         .take(100000)
